Add trivial store, policy language and extra data.
Implement a very simple store, with buckets and objects, and allow the setting
of simple policies on buckets.

This allows us to issue queries that get different results based on the
bucket and policy. The policy doesn't extend to individual objects yet, let's
see if that's even necessary for what we need this program to do.

The policy language is very simple, and is just a JSON document with an
'action' (allow or deny) and a 'requires' that allows specification of whether
or not extra data fields are present.