Auto-create buckets to match test script.
The test script now creates buckets with time-specific names, and they have
standard suffixes, '1', '1v', '1l'. We care about the number here as it
dictates what policy a bucket is expected to have.

Currently, 1 indicates 'regular', 2 indicates 'regular but require extra
data', 3 indicates 'always deny', and 4 indicates 'create an extra data loop
to make sure this error is handled'. We extract the number and create the
bucket with the appropriate policy.

Not a lot of bounds checking going on at the moment.